# Optimize `formatToData` for large numbers