**What**

- Replaced O(n²) unshift operations with O(n) push + reverse for building the integer parts array, improving scalability for large numbers.

- Switched string concatenation (+=) to an array-based join('') for valueAsString to minimize reallocations.


**Benchmarks**

Run on Node.js v22.11.0 (Apple M3, ~3.55 GHz)

- Small (5 digits): Old 1.76 µs/iter, New 1.71 µs/iter (~3% faster)
- Medium (8 digits): Old 2.43 µs/iter, New 2.33 µs/iter (~4% faster)
- Large (12 digits): Old 3.13 µs/iter, New 2.96 µs/iter (~5.4% faster)
- Very Large (15 digits): Old 3.77 µs/iter, New 3.54 µs/iter (~6.1% faster)
- Extreme (18 digits): Old 4.44 µs/iter, New 4.15 µs/iter (~6.5% faster)